requirements.txt
discord.py==1.6.0
PyNaCl==1.3.0
dnspython==1.16.0
async-timeout==3.0.1
requests
youtube_dl
os